+++ WTFPlugin/src/wtfplugin/actions/LaunchTomcat9.java
@@ -42,17 +42,17 @@ import org.eclipse.ui.actions.ActionDelegate;
 import wtfplugin.Activator;
 import wtfplugin.preferences.WTFPreferences;
 
-public class LaunchTomcat7 {
+public class LaunchTomcat9 {
 	
 
-	public LaunchTomcat7() {
+	public LaunchTomcat9() {
 	}
 
 
 	public static void launchTomcat(IProject project, String webappname, String port, String httpsPort, String serverPort, String contextContent, String jvmArgs) throws CoreException, IOException, JavaModelException {
 		
 		IProcess process= DebugUITools.getCurrentProcess();
-		if (process != null && process.getLaunch().getLaunchConfiguration().getName().equals(LaunchTomcat9.getLaunchName(project))) {
+		if (process != null && process.getLaunch().getLaunchConfiguration().getName().equals(getLaunchName(project))) {
 			process.terminate();
 		}
 			
@@ -63,22 +63,22 @@ public class LaunchTomcat7 {
 		ILaunchConfiguration[] configurations = manager.getLaunchConfigurations(type);
 		for (int i = 0; i < configurations.length; i++) {
 			ILaunchConfiguration configuration = configurations[i];
-			if (configuration.getName().equals(LaunchTomcat9.getLaunchName(project))) {
+			if (configuration.getName().equals(getLaunchName(project))) {
 				configuration.delete();
 				break;
 			}
 		}
 		IVMInstall jre = JavaRuntime.getDefaultVMInstall();
 		File jdkHome = jre.getInstallLocation();
-		ILaunchConfigurationWorkingCopy workingCopy = type.newInstance(null, LaunchTomcat9.getLaunchName(project));
+		ILaunchConfigurationWorkingCopy workingCopy = type.newInstance(null, getLaunchName(project));
 		workingCopy.setAttribute(IJavaLaunchConfigurationConstants.ATTR_VM_INSTALL_NAME, jre.getName());
 		workingCopy.setAttribute(IJavaLaunchConfigurationConstants.ATTR_VM_INSTALL_TYPE, jre.getVMInstallType().getId());
 		workingCopy.setAttribute(IJavaLaunchConfigurationConstants.ATTR_MAIN_TYPE_NAME, "org.apache.catalina.startup.Bootstrap");
-		workingCopy.setAttribute(IJavaLaunchConfigurationConstants.ATTR_PROGRAM_ARGUMENTS, "-config "+System.getProperty("java.io.tmpdir") + "/"+LaunchTomcat9.getLaunchName(project)+"_server.xml"+" start");
+		workingCopy.setAttribute(IJavaLaunchConfigurationConstants.ATTR_PROGRAM_ARGUMENTS, "-config "+System.getProperty("java.io.tmpdir") + "/"+getLaunchName(project)+"_server.xml"+" start");
 		IPath toolsPath = new Path(jdkHome.getAbsolutePath()).append("lib").append("tools.jar");
 		IRuntimeClasspathEntry toolsEntry = JavaRuntime.newArchiveRuntimeClasspathEntry(toolsPath);
 		toolsEntry.setClasspathProperty(IRuntimeClasspathEntry.USER_CLASSES);
-		IPath bootstrapPath = new Path(WTFPreferences.getTomcatVariable("7")).append("bin").append("bootstrap.jar");
+		IPath bootstrapPath = new Path(WTFPreferences.getTomcatVariable("9")).append("bin").append("bootstrap.jar");
 		IRuntimeClasspathEntry bootstrapEntry = JavaRuntime.newVariableRuntimeClasspathEntry(bootstrapPath);
 		bootstrapEntry.setClasspathProperty(IRuntimeClasspathEntry.USER_CLASSES);
 		IPath systemLibsPath = new Path(JavaRuntime.JRE_CONTAINER);
@@ -90,7 +90,7 @@ public class LaunchTomcat7 {
 
 		addTomcatBinJarToClasspath(classpath, "tomcat-juli.jar");
 		
-		File tomcatLib = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("7")).append("lib").toFile();
+		File tomcatLib = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("9")).append("lib").toFile();
 		for (String jarFile : tomcatLib.list()) {
 			if (jarFile.endsWith(".jar")) {
 				addTomcatLibJarToClasspath(classpath, jarFile);
@@ -114,7 +114,7 @@ public class LaunchTomcat7 {
 		// Borro el keystore viejo
 		org.apache.commons.io.FileUtils.deleteQuietly(new File(System.getProperty("user.home") + "/.keystoreWTF"));
 		if (LaunchTomcat.keystore == null) {
-			InputStream is = LaunchTomcat7.class.getResourceAsStream("keystoreWTF");
+			InputStream is = LaunchTomcat9.class.getResourceAsStream("keystoreWTF");
 			LaunchTomcat.keystore = IOUtils.toByteArray(is);
 			is.close();
 		}
@@ -122,11 +122,12 @@ public class LaunchTomcat7 {
 		org.apache.commons.io.FileUtils.writeByteArrayToFile(keystoreFile, LaunchTomcat.keystore);
 		
 		// Excribo el context.xml
-		File contextDir = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("7")).toFile();
-		if (WTFPreferences.clusterSessionMongo()) {
+		File contextDir = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("9")).toFile();
+		// no esta soportado por el momento
+		if (WTFPreferences.clusterSessionMongo() && false) {
 			org.apache.commons.io.FileUtils.deleteQuietly(new File(contextDir + "/conf/context.xml"));
 			if (LaunchTomcat.contextxml== null) {
-				InputStream is = LaunchTomcat7.class.getResourceAsStream("context.xml");
+				InputStream is = LaunchTomcat9.class.getResourceAsStream("context.xml");
 				String tempServer = IOUtils.toString(is);
 				LaunchTomcat.contextxml = tempServer;
 				is.close();
@@ -138,10 +139,10 @@ public class LaunchTomcat7 {
 			org.apache.commons.io.FileUtils.writeStringToFile(fileContext, newTempContext);
 		}
 		// Primero escribo el server xml
-		org.apache.commons.io.FileUtils.deleteQuietly(new File(System.getProperty("java.io.tmpdir") + "/"+LaunchTomcat9.getLaunchName(project)+"_server.xml"));
+		org.apache.commons.io.FileUtils.deleteQuietly(new File(System.getProperty("java.io.tmpdir") + "/"+project.getName()+"_server.xml"));
 //		if (serverxml == null) {
 		{
-			InputStream is = LaunchTomcat7.class.getResourceAsStream("server.xml");
+			InputStream is = LaunchTomcat9.class.getResourceAsStream("server.xml");
 			String tempServer = IOUtils.toString(is);
 			LaunchTomcat.serverxml = tempServer;
 			is.close();
@@ -166,25 +167,34 @@ public class LaunchTomcat7 {
 		newTempServer = newTempServer.replace("@workdir@", workdir);
 		newTempServer = newTempServer.replace("@contextcontent@", contextContent);	
 		
-		File f = new File(System.getProperty("java.io.tmpdir") + "/"+LaunchTomcat9.getLaunchName(project)+"_server.xml");
+		File f = new File(System.getProperty("java.io.tmpdir") + "/" + getLaunchName(project) + "_server.xml");
 		org.apache.commons.io.FileUtils.writeStringToFile(f, newTempServer);
 		
 		// copio las librerias de mongo si no estan
-		File mongoStore = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("7")).append("lib").append("mongo-store-proxy-1.7.3.jar").toFile();
-		if (!mongoStore.exists()) {
-			InputStream is = LaunchTomcat.class.getResourceAsStream("mongo-store-proxy-1.7.3.jar");
-			FileOutputStream fout= new FileOutputStream(mongoStore);
-			IOUtils.copy(is, fout);
-			is.close();
-			fout.close();
-		}
-		File mongoDriver = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("7")).append("lib").append("mongo-java-driver-3.2.1.jar").toFile();
-		if (!mongoDriver.exists()) {
-			InputStream is = LaunchTomcat7.class.getResourceAsStream("mongo-java-driver-3.2.1.jar");
-			FileOutputStream fout= new FileOutputStream(mongoDriver);
-			IOUtils.copy(is, fout);
-			is.close();
-			fout.close();
+		File mongoStore = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("9")).append("lib").append("mongo-store-proxy-1.9.jar").toFile();
+		File mongoDriver = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("9")).append("lib").append("mongo-java-driver-3.2.1.jar").toFile();
+		if (WTFPreferences.clusterSessionMongo() && false) {
+			if (!mongoStore.exists()) {
+				InputStream is = LaunchTomcat.class.getResourceAsStream("mongo-store-proxy-1.9.jar");
+				FileOutputStream fout= new FileOutputStream(mongoStore);
+				IOUtils.copy(is, fout);
+				is.close();
+				fout.close();
+			}
+			if (!mongoDriver.exists()) {
+				InputStream is = LaunchTomcat9.class.getResourceAsStream("mongo-java-driver-3.2.1.jar");
+				FileOutputStream fout= new FileOutputStream(mongoDriver);
+				IOUtils.copy(is, fout);
+				is.close();
+				fout.close();
+			}
+		} else {
+			if (mongoStore.exists()) {
+				mongoStore.delete();
+			}
+			if (mongoDriver.exists()) {
+				mongoDriver.delete();
+			}
 		}
 		
 		IJavaProject javaProject = JavaCore.create(project);
@@ -213,11 +223,16 @@ public class LaunchTomcat7 {
 				+ "-Dcatalina.base=\"..\"" + "-Dcatalina.home=\"..\"" + "-Djava.io.tmpdir=\"..\\temp\" " + (jvmArgsParams));
 		
 		
-		File workingDir = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("7")).append("bin").toFile();
+		File workingDir = JavaCore.getClasspathVariable(WTFPreferences.getTomcatVariable("9")).append("bin").toFile();
 		workingCopy.setAttribute(IJavaLaunchConfigurationConstants.ATTR_WORKING_DIRECTORY, workingDir.getAbsolutePath());
 		ILaunchConfiguration configuration = workingCopy.doSave();
 		DebugUITools.launch(configuration, ILaunchManager.DEBUG_MODE);
 	}
+
+	public static String getLaunchName(IProject project) {
+		return "start_" + project.getName();
+	}
+
 
 	public static void addToClasspath( List classpath, IJavaProject javaProject) throws JavaModelException, CoreException {
 		IClasspathEntry[] entries = javaProject.getRawClasspath(); 
@@ -239,13 +254,13 @@ public class LaunchTomcat7 {
 	}
 	
 	public static void addTomcatBinJarToClasspath(List classpath, String jar) throws CoreException {
-		IPath jspapijar = new Path(WTFPreferences.getTomcatVariable("7")).append("bin").append(jar);
+		IPath jspapijar = new Path(WTFPreferences.getTomcatVariable("9")).append("bin").append(jar);
 		IRuntimeClasspathEntry jspapijarEntry = JavaRuntime.newVariableRuntimeClasspathEntry(jspapijar);
 		classpath.add(jspapijarEntry.getMemento());
 	}
 	
 	public static void addTomcatLibJarToClasspath(List classpath, String jar) throws CoreException {
-		IPath jspapijar = new Path(WTFPreferences.getTomcatVariable("7")).append("lib").append(jar);
+		IPath jspapijar = new Path(WTFPreferences.getTomcatVariable("9")).append("lib").append(jar);
 		IRuntimeClasspathEntry jspapijarEntry = JavaRuntime.newVariableRuntimeClasspathEntry(jspapijar);
 		classpath.add(jspapijarEntry.getMemento());
 	}
